+++ docs/components/examples/ExStyling-1.vue
@@ -3,12 +3,7 @@
     mode='single'
     tint-color='#ff6666'
     v-model='selectedDate'
-    :style='wrapperStyle'
-    :header-style='headerStyle'
-    :header-divider-color='headerDividerColor'
-    :weekday-style='weekdayStyle'
-    :weeks-style='weeksStyle'
-    :weekday-divider-color='headerDividerColor'
+    :theme-styles='themeStyles'
     is-double-paned
     is-inline>
   </v-date-picker>
@@ -19,25 +14,32 @@ export default {
   data() {
     return {
       selectedDate: null,
-      wrapperStyle: {
-        border: '0',
-      },
-      headerStyle: {
-        color: '#fafafa',
-        backgroundColor: '#4a5867',
-        border: 'solid #404c59',
-        borderWidth: '1px 1px 0 1px',
-      },
-      headerDividerColor: '#404c59',
-      weekdayStyle: {
-        color: '#98aeb3',
-        backgroundColor: '#4a5867',
-        border: 'solid #384763',
-        borderWidth: '0 1px',
-        padding: '5px 0 10px 0',
-      },
-      weeksStyle: {
-        border: '1px solid #dadada',
+      themeStyles: {
+        wrapper: {
+          border: '0',
+        },
+        header: {
+          color: '#fafafa',
+          backgroundColor: '#4a5867',
+          borderColor: '#404c59',
+          borderWidth: '1px 1px 0 1px',
+        },
+        headerVerticalDivider: {
+          borderLeft: '1px solid #404c59',
+        },
+        weekdays: {
+          color: '#98aeb3',
+          backgroundColor: '#4a5867',
+          borderColor: '#384763',
+          borderWidth: '0 1px',
+          padding: '5px 0 10px 0',
+        },
+        weekdaysVerticalDivider: {
+          borderLeft: '1px solid #404c59',
+        },
+        weeks: {
+          border: '1px solid #dadada',
+        },
       },
     };
   },