+++ docs/components/pages/PageStyling.vue
@@ -14,6 +14,8 @@
 <script>
 import ExStyling1 from '../examples/ExStyling-1';
 import ExStyling1Code from '!!raw-loader!../examples/ExStyling-1';
+import ExStyling2 from '../examples/ExStyling-2';
+import ExStyling2Code from '!!raw-loader!../examples/ExStyling-2';
 
 import SectionStyling from '../sections/SectionStyling';
 
@@ -28,6 +30,11 @@ export default {
           title: 'Styled Header & Weekday Sections',
           code: ExStyling1Code,
           component: ExStyling1,
+        },
+        {
+          title: 'Example 2',
+          code: ExStyling2Code,
+          component: ExStyling2,
         },
       ],
     };