+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 0.0.7
+* Add 'theme-styles' prop to calendar for theming
+* Remove a number of '...-style' replaced by theme-styles object
+* Renamed 'select-mode' prop to 'mode' for datepicker
+
+## 0.0.6
+* Fix README
+
 ## 0.0.5
 
 * Add tap to dismiss popover
-* Add color blending for days not in month (transparent layers are visually buggy and not as performant)
 * Add tint color (with blending) for easy date picker styling
 * Add support for custom title position (left, center, right)